app: move GET /api/release to app.routers.release
